+++ src/EulerHS/Framework/Flow/Interpreter.hs
@@ -33,6 +33,7 @@ import           EulerHS.Prelude
 import qualified Network.Connection as Conn
 import qualified Network.HTTP.Client as HTTP
 import qualified Network.HTTP.Client.TLS as TLS
+import           Network.HTTP.Client.Internal
 import qualified Network.HTTP.Types as HTTP
 import qualified Network.TLS as TLS
 import qualified Network.TLS.Extra.Cipher as TLS
@@ -190,8 +191,11 @@ interpretFlowMethod mbFlowGuid flowRt@R.FlowRuntime {..} (L.CallServantAPI mbMgr
       case mbClientMngr of
         Right mngr -> do
           let S.ClientEnv manager baseUrl cookieJar makeClientRequest = S.mkClientEnv mngr bUrl
+          let setR req = if HTTP.responseTimeout req == HTTP.responseTimeoutNone
+                            then setRequestTimeout T.defaultTimeout req
+                            else req {HTTP.responseTimeout = mResponseTimeout mngr}
           eitherResult <- S.runClientM (T.runEulerClient (dbgLogger T.Debug) getLoggerMaskConfig bUrl clientAct) $
-            S.ClientEnv manager baseUrl cookieJar (\url -> setRequestTimeout T.defaultTimeout . makeClientRequest url)
+            S.ClientEnv manager baseUrl cookieJar (\url -> setR . makeClientRequest url)
           case eitherResult of
             Left err -> do
               dbgLogger T.Error $ show err
@@ -210,7 +214,7 @@ interpretFlowMethod mbFlowGuid flowRt@R.FlowRuntime {..} (L.CallServantAPI mbMgr
     getLoggerMaskConfig =
       R.getLogMaskingConfig . R._loggerRuntime . R._coreRuntime $ flowRt
 
-interpretFlowMethod _ R.FlowRuntime {..} (L.CallHTTP request cert next) =
+interpretFlowMethod _ flowRt@R.FlowRuntime {..} (L.CallHTTP request cert next) =
     fmap next $ do
       httpLibRequest <- getHttpLibRequest request
       _manager <- maybe (pure $ Right _defaultHttpClientManager) mkManagerFromCert cert
@@ -228,9 +232,25 @@ interpretFlowMethod _ R.FlowRuntime {..} (L.CallHTTP request cert next) =
             Right httpResponse -> do
               case translateHttpResponse httpResponse of
                 Left errMsg -> do
+                  logJsonError errMsg (T.maskHTTPRequest getLoggerMaskConfig request)
                   pure $ Left errMsg
                 Right response -> do
+                  logJson T.Debug 
+                    $ T.HTTPRequestResponse 
+                      (T.maskHTTPRequest getLoggerMaskConfig request) 
+                      (T.maskHTTPResponse getLoggerMaskConfig response)
                   pure $ Right response
+  where
+    logJsonError :: Text -> T.HTTPRequest -> IO ()
+    logJsonError err = logJson T.Error . T.HTTPIOException err
+    logJson :: ToJSON a => T.LogLevel -> a -> IO ()
+    logJson debugLevel =
+      R.runLogger (Just "API CALL:") (R._loggerRuntime . R._coreRuntime $ flowRt)
+        . L.logMessage' debugLevel ("callHTTP" :: String)
+        . encodeJSON
+    
+    getLoggerMaskConfig = 
+      R.getLogMaskingConfig . R._loggerRuntime . R._coreRuntime $ flowRt
 
 interpretFlowMethod mbFlowGuid R.FlowRuntime {..} (L.EvalLogger loggerAct next) =
   next <$> R.runLogger mbFlowGuid (R._loggerRuntime _coreRuntime) loggerAct