+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog for euler-hs
 
+## [2.0.4.4] - 2021-04-13
+
+* Rename `changeLoggerContext` to `withLoggerContext`
+* Add `updateLoggerContext`
+
+## [2.0.4.3] - 2021-04-12
+
+* Add `changeLoggerContext` to EulerHS.Extra.Language
+
+## [2.0.4.2] - 2021-04-07
+
+* Fixed a bug with MVar blocking when disposing `Async` logger.
+
 ## [2.0.4.1] - 2021-03-03
 
 * Remove waste entries.